refactor(srv): rename storage module memory_bundles -> bundles
Phase 1 step A of SPEC_ARMORY_NAMING_CONSOLIDATION_2026_09_09.md. The module
holds db_bundles CRUD; 'memory' in its name predates the Preset->Bundle
rename. Pure rename via git mv + path-reference sweep; no behaviour change.

The legacy migration literals "db_memory_bundles", "idx_memory_bundles_is_blank"
and "idx_ss_memory_bundles_is_blank" in migrations.rs are deliberately
untouched — they are adoption entries for tables that must keep their old
names forever. Doc-comment mentions of db_memory_bundles describing migration
history are likewise left as-is.
