+++ agentmux-srv/src/backend/storage/mod.rs
@@ -25,7 +25,7 @@ pub mod history;
 pub mod identities;
 pub mod managed;
 pub mod mcp_servers;
-pub mod memory_bundles;
+pub mod bundles;
 pub mod migrations;
 pub mod muxbus;
 pub mod registry_mirror;
@@ -45,5 +45,5 @@ pub use error::StoreError;
 pub use history::AgentHistory;
 pub use identities::{AgentIdentityLink, IdentityAccount, SecretRef};
 pub use mcp_servers::McpServer;
-pub use memory_bundles::{format_global_brain_block, Memory};
+pub use bundles::{format_global_brain_block, Memory};
 pub use skills::{AgentSkill, Skill};
